url_api/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apturl==0.5.2
arcade==2.6.16
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
CacheControl==0.12.11
cachetools==5.3.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
cloudinary==1.29.0
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dataclasses-json==0.5.7
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dj-database-url==1.2.0
Django==4.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-jet-reboot==1.3.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
firebase-admin==6.1.0
fonttools==4.37.4
frozenlist==1.3.3
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gpt-index==0.4.28
greenlet==2.0.2
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
gyp==0.1
html2text==2020.1.16
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.4.4
langchain==0.0.109
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llama-index==0.4.28
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.6.0
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==1.23.3
oauthlib==3.2.0
olefile==0.46
openai==0.27.2
packaging==21.3
pandas==1.5.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.1.1
platformdirs==2.5.1
proto-plus==1.22.2
protobuf==4.22.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.6
pygame==2.1.2
pyglet==2.0.dev23
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.6.0
pymacaroons==0.13.0
pymunk==6.2.1
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqtgraph==0.13.1
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python3-openid==3.2.0
pytiled-parser==2.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
regex==2022.10.31
reportlab==3.6.8
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
SQLAlchemy==1.4.46
sqlparse==0.4.2
ssh-import-id==5.11
stripe==4.0.2
systemd-python==234
tenacity==8.2.2
tiktoken==0.3.1
tkintertable==1.3.3
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.4.0
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
whitenoise==6.2.0
xdg==5
xkit==0.0.0
yarl==1.8.2
zipp==1.0.0
